// src/store/booking/types.ts
export interface BookingProps {
  bookingState: BookingState;
  codeDiscount?: string;
  flexiType?: FlexiType;
  fromStationId?: number;
  lineGroupCode?: string;
  lineNumber?: number;
  passengers?: PassengerData[];
  percentualDiscountIds?: number[];
  price?: number;
  discountAmount?: number;
  priceSource?: string;
  routeId?: RouteId;
  seatClass?: SeatClass;
  sections?: SectionWithSeats[];
  selectedAddons?: SelectedAddon[];
  tariffs?: Tariff[];
  toStationId?: number;
  type?: TicketType;
}

export interface Booking {
  back: BookingProps | null;
  connection: Partial<ConnectionPayload>;
  isReturn: boolean;
  there: BookingProps | null;
}

export interface BookingReducerProps {
  booking: Booking | null;
  fav: ConnectionLocations[];
}